modify addservice

+++ MechanicsMateBackend/Models/VM/VehicleData.cs
@@ -9,7 +9,7 @@ namespace MechanicsMateBackend.Models.VM
     public class VehicleData
     {
         public int OwnerId { get; set; }
-        public ulong VehicleInfoId { get; set; }
+        public int VehicleInfoId { get; set; }
         public uint Mileage { get; set; }
         public uint? DrivingHabit { get; set; }
     }
